Move utility functions to infodialog/utils.py
